update db columns
