+++ scraper/src/merge_datasets.js
@@ -2,44 +2,7 @@ import fs from 'fs';
 import { parse as json2csv } from 'json2csv';
 import csv from 'csvtojson';
 import permapeopleColumnMapping from './helpers/column_mapping_permapeople.js';
-import { sanitizeColumnNames } from './helpers/helpers.js';
-
-function getSoilPH(pH) {
-  // If pH is a range, split it into two numbers and calculate the average
-  if (pH.includes('-')) {
-    const [min, max] = pH.split('-').map(Number);
-    pH = (min + max) / 2;
-  } else if (pH.includes('–')) {
-    // Handle em dash character
-    const [min, max] = pH.split('–').map(Number);
-    pH = (min + max) / 2;
-  } else if (pH.includes('<')) {
-    // Handle less than symbol
-    pH = Number(pH.slice(1)) - 0.1;
-  } else if (pH.includes('>')) {
-    // Handle greater than symbol
-    pH = Number(pH.slice(1)) + 0.1;
-  } else {
-    pH = Number(pH);
-  }
-
-  if (isNaN(pH) || pH < 0 || pH > 14) {
-    // Handle invalid pH values
-    return null;
-  } else if (pH <= 5.0) {
-    return 'very acid';
-  } else if (pH >= 5.1 && pH <= 6.5) {
-    return 'acid';
-  } else if (pH >= 6.6 && pH <= 7.3) {
-    return 'neutral';
-  } else if (pH >= 7.4 && pH <= 7.8) {
-    return 'alkaline';
-  } else if (pH >= 7.9) {
-    return 'very alkaline';
-  } else {
-    return null;
-  }
-}
+import { sanitizeColumnNames, getSoilPH } from './helpers/helpers.js';
 
 const unifyValueFormat = (plants, columnMapping) => {
   const mappedColumns = Object.keys(columnMapping).filter((key) => columnMapping[key] !== null);
@@ -150,7 +113,10 @@ async function mergeDatasets() {
 
   practicalPlants.forEach((plant) => {
     const binomial_name = plant['binomial_name'];
-    const plantInPermapeople = permapeople.find((plant) => plant.scientific_name === binomial_name);
+    const plantInPermapeople = permapeople.find(
+      (plant) => plant['scientific_name'] === binomial_name,
+    );
+
     if (plantInPermapeople) {
       const mergedPlant = {};
       Object.keys(plantInPermapeople).forEach((key) => {
@@ -181,7 +147,7 @@ async function mergeDatasets() {
   permapeople.forEach((plant) => {
     const scientific_name = plant['scientific_name'];
     const plantInPracticalPlants = practicalPlants.find(
-      (plant) => plant.scientific_name === scientific_name,
+      (plant) => plant['scientific_name'] === scientific_name,
     );
     if (!plantInPracticalPlants) {
       allPlants.push({
